Enemy collision with explosion + ConcurrentModificationException fix
I fixed an issue where the app would crash because the ArrayList enemigos was accessed in render while being modified (removing an enemy) in update. The app now uses a for loop instead of a for-each to avoid this

I also added collisions with enemies, which now explode when touching the player once.

The explosions reproduce an animation and when they finished they are removed