Leaderboards top 5, points during game + BlackScreen bug fix, cars and sounds additions
I added the points feature. During the game you get points, extra points when picking up coins and lives, penalty when crashing with other cars

The top scores are stored with the nickname that got them. The leaderboard dialog shows the top 5, which is updated when finishing a game, adding the new score with the current nickname

The user now can also change the nickname touching the nick text

The coin icon in the title screen reproduces the coin sound when touched

I added two new cars and two explosion sounds, and a car start sound when starting a game

I changed the leaderboards icon from the medal to the podium

I fixed a bug when going start, back, start back, removing the callbacks in BlackScreenActivity to avoid starting more than 1 game